refactor: TableGroup 패키지 분리
