+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -13,7 +13,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
-import kitchenpos.domain.BaseEntity;
+import suppoert.domain.BaseEntity;
 
 @Entity
 @Table(name = "orders")
