+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -5,7 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import kitchenpos.domain.BaseEntity;
+import suppoert.domain.BaseEntity;
 import kitchenpos.menu.domain.Menu;
 
 @Entity
