+++ src/main/java/kitchenpos/tablegroup/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -7,9 +7,10 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.OneToMany;
 import kitchenpos.order.domain.OrderTable;
+import suppoert.domain.BaseEntity;
 
 @Entity
-public class TableGroup extends BaseEntity{
+public class TableGroup extends BaseEntity {
 
     @Column(nullable = false)
     private LocalDateTime createdDate = LocalDateTime.now();
