+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.tablegroup.domain;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
